Protect post controller

+++ blog-app/app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller {
     public function index(){
@@ -31,7 +32,9 @@ class PostController extends Controller {
 
     // create logic
     public function store(Request $request){
-        $validData = $request->validate([
+
+        if (Gate::allows('is-active')) {
+            $validData = $request->validate([
             'title' => 'required',
             'tags' => 'required',
             'content' => 'required',
@@ -45,58 +48,67 @@ class PostController extends Controller {
             'photo.mimes' => 'Prześlij obraz w formacie JPEG lub PNG.',
             'photo.max' => 'Maksymalny rozmiar obrazu to 2 MB.',
         ]);
-        $validData['photo']=$request->file('photo')->store('logos', 'public');
+            $validData['photo']=$request->file('photo')->store('logos', 'public');
 
-        // Create a new post instance
-        $post = new Post;
-        $post->title = $validData['title'];
-        $post->tags = $validData['tags'];
-        $post->content = $validData['content'];
-        $post->photo = $validData['photo'];
-        $post->date = now();
-        $post->user_id = auth()->id();
-        $post->score = 8;
+            // Create a new post instance
+            $post = new Post;
+            $post->title = $validData['title'];
+            $post->tags = $validData['tags'];
+            $post->content = $validData['content'];
+            $post->photo = $validData['photo'];
+            $post->date = now();
+            $post->user_id = auth()->id();
+            $post->score = 8;
+
+            $post->save();
 
 
-        $post->save();
+            // Increase user score
+            $user = User::find(auth()->id());
+            $user->score += $post->score;
+            $user->update();
 
-
-        // Increase user score
-        $user = User::find(auth()->id());
-        $user->score += $post->score;
-        $user->update();
-
-        $rankings = [1, 2, 3, 4, 5];
-        foreach($rankings as $rank){
-            if($user->score >= $rank * 20){
-                $user->rankings()->syncWithoutDetaching([$rank]);
+            $rankings = [1, 2, 3, 4, 5];
+            foreach($rankings as $rank){
+                if($user->score >= $rank * 20){
+                    $user->rankings()->syncWithoutDetaching([$rank]);
+                }
             }
-        }
-        $message = [
-            'content' => "Post został utworzony",
-            'type' => 'success'
-        ];
+            $message = [
+                'content' => "Post został utworzony",
+                'type' => 'success'
+            ];
 
-        return Redirect('/')->with('message', $message);
+            return Redirect('/')->with('message', $message);
+        } else {
+            abort(403);
+        }
     }
 
     public function delete(Post $post){
-        $post->delete();
-        $message = [
-            'content' => "Post został usunięty",
-            'type' => 'delete'
-        ];
-        return redirect('/')->with('message', $message);
+        if (Gate::allows('post-operation', $post)) {
+            $post->delete();
+            $message = [
+                'content' => "Post został usunięty",
+                'type' => 'delete'
+            ];
+            return redirect('/')->with('message', $message);
+        } else {
+            abort(403);
+        }
     }
 
     public function edit(Post $post){
-        return view('posts.edit', [
-        "post"=> $post
-        ]);
+        if (Gate::allows('post-operation', $post)) {
+            return view('posts.edit', ["post"=>$post]);
+        } else {
+            abort(403);
+        }
     }
 
     public function update(Request $request, Post $post){
-        $formFields = $request->validate([
+        if (Gate::allows('post-operation', $post)) {
+            $formFields = $request->validate([
             'title'=>'required',
             'tags'=>'required',
             'content'=>'required',
@@ -113,5 +125,8 @@ class PostController extends Controller {
         ];
 
         return redirect('/posts/'.$post->id)->with('message', $message);
+        } else {
+            abort(403);
+        }
     }
 }
